+++ src/main/java/com/example/api_spring/repositories/EsporteRepository.java
@@ -3,5 +3,5 @@ package com.example.api_spring.repositories;
 import com.example.api_spring.models.Esporte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EsporteRepository extends JpaRepository<Esporte,Integer> {
+public interface EsporteRepository extends JpaRepository<Esporte,Long> {
 }
